Heap-buffer-overflow in void std::__final_insertion_sort.
It was getting crash while sorting rowspan cells. std::sort calls comparator
function 2 times with same value in reverse order when first call return
true for verification of result from second call which should be compulsory
false. But it was not there in our comparator function.

There was a mistake in comparator function. Cells should be sorted based
on row indexes only when one cell is not a subpart of another cell.

R=jchaffraix@chromium.org
BUG=296003

Review URL: https://chromiumcodereview.appspot.com/23874023

git-svn-id: svn://svn.chromium.org/blink/trunk@158347 bbb929c8-8fbe-4397-9dbb-9b2b20218538