Do not layout for position changes in RenderWidget::setWidgetGeometry(...)
Previously, a layout could be caused by a RenderWidget changing
position. For example, an iframe in a scrollable area would get called
to layout if the area scrolled. This occurred because of the the following
callstack:
FrameView calls
RenderWidget::updateWidgetPosition(...) which calls
RenderWidget::updateWidgetGeometry(...) which calls
RenderWidget::setWidgetGeometry(...)

setWidgetGeometry would return true and force a layout if the frame rect
changed position. This should only return true if the frame rect actually
resizes, so this patch updates setWidgetGeometry to return false (not
forcing a layout) if just the frame's position changes.

BUG=300275
TEST=manual, verified using the web inspector that additional layouts
are not performed when scrolling iframes.

Review URL: https://chromiumcodereview.appspot.com/25154002

git-svn-id: svn://svn.chromium.org/blink/trunk@158500 bbb929c8-8fbe-4397-9dbb-9b2b20218538