blink_perf/Layout_fixed-grid-lots-of-data regression on Mac
The regression comes from calling RenderBox::hasRelativeLogicalHeight,
which is virtual, if no size changed like in the test case. The fix is
to change the order in which we evaluate the relayout check.

This solves the regression from the original change on my machine.

BUG=290989

Review URL: https://codereview.chromium.org/25402002

git-svn-id: svn://svn.chromium.org/blink/trunk@158593 bbb929c8-8fbe-4397-9dbb-9b2b20218538