Fix 1-pixel decodedSize mismatch of downsampled JPEG
decodedSize() needs to exactly match the output_width/output_height of
the JPEG decoder. The original code may result a 1-pixel mismatch of
the decoded size if the original width/height is not multiple of 8,
causing data corruption when the image buffer is used to create a
sub-image, and crashes when the sub-image is used.

BUG=276666,301973
TEST=JPEGImageDecoderTest.downsampleImageSizeMultipleOf8, JPEGImageDecoderTest.downsampleImageSizeNotMultipleOf8

Review URL: https://codereview.chromium.org/25382002

git-svn-id: svn://svn.chromium.org/blink/trunk@158615 bbb929c8-8fbe-4397-9dbb-9b2b20218538