RenderBox::paintFillLayers should check box-shadow when finding  an opaque layer.
RenderBox::paintFillLayers has a fast path when a layer has an opaque background image.
If a layer has an opaque background image, basically the layer's background layers will be hidden by the background image. So we can skip rendering the background layers.
However, if box-shadow is also specified, the box-shadow will be rendered in a skipped backround layer.

This causes "box-shadow disappears". We cannot use the fast path if box-shadow is specified.

BUG=260645
TEST=fast/box-shadow/normal-box-shadow-with-background-image.html

Review URL: https://codereview.chromium.org/23143008

git-svn-id: svn://svn.chromium.org/blink/trunk@158643 bbb929c8-8fbe-4397-9dbb-9b2b20218538