document.documentElement.scrollTop/Left is zero
Blink, as of today, is the only engine that does not respect the
specification with regards to the values returned from scrollTop
and scrollLeft properties of documentElement and body nodes.

Current behavior is: in both strict and quirks modes, documentElement.scrollTop
always returns 0, while body.scrollTop always returns scrollY.
The analogy behavior applies to scrollLeft and scrollX (the return).

This patch fixes Blink's behavior in accordance to specification [1]:
if a document is in strict mode, document.documentElement.scrollTop
should return 'scrollY', and document.body.scrollTop should return 0.
On the other hand, if it is in quirks mode, document.documentElement
should return 0, and document.body.scrollTop should return 'scrollY'.
Same analogy applies to scrollLeft property.

An optimistic approach was chosen to fix this in the sense of doing this
change without collecting usage data first. The rationale behind it is
that given that Blink is the only engine behavior'ing differently, relevant
WebSites would already have two codepaths.

[1] http://dev.w3.org/csswg/cssom-view/#dom-element-scrolltop

Test: fast/dom/Element/scrollTop-scrollLeft-strict-quirks-modes.html
Existing updated tests:
      fast/dom/anchor-without-content.html
      fast/events/mouse-cursor-image-set.html
      fast/events/mouse-cursor.html
      fast/events/touch/gesture/touch-gesture-fully-scrolled-iframe-propagates.html
      fast/events/touch/gesture/touch-gesture-noscroll-body-propagated.html
      fast/events/touch/gesture/touch-gesture-noscroll-body-xhidden.html
      fast/events/touch/gesture/touch-gesture-noscroll-body-yhidden.html
      fast/events/touch/gesture/touch-gesture-scroll-iframe-editable.html
      fast/events/touch/gesture/touch-gesture-scroll-iframe-not-propagated.html
      fast/events/touch/gesture/touch-gesture-scroll-page-propagated.html
      fast/events/touch/gesture/touch-gesture-scroll-page.html
      fast/scrolling/hover-during-scroll.html
      http/tests/navigation/resources/frame-with-anchor-same-origin.html
      http/tests/navigation/resources/frame-with-anchor.html

BUG=157855

Review URL: https://codereview.chromium.org/25039006

git-svn-id: svn://svn.chromium.org/blink/trunk@158719 bbb929c8-8fbe-4397-9dbb-9b2b20218538