Update baselines after r159046.
It looks like this test was already (accidentally) testing
the issue which r159046 is fixing by disabling this (broken) optimization.

It looks like when we zoom we don't necessarily always fit exactly to the
window, but may be off by as much as a pixel.  I'm not sure if that's a
bug or not, but we certainly should be reporting that correctly to JS.

BUG=290399
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/26298005

git-svn-id: svn://svn.chromium.org/blink/trunk@159052 bbb929c8-8fbe-4397-9dbb-9b2b20218538