[CSS Blending] Implement mix-blend-mode in software. The implementation follows the CSS Blending and compositing spec at http://dev.w3.org/fxtf/compositing-1/
It seems that there might be slightly better performance if we use the software path for CSS blending. Basically, on the hardware path, there are some bitmap copy operations involved, and we'd like to avoid this whenever possible.

The hardware implementation, currently in the works will only be followed for elements that force the creation of a composited layer.

BUG=288655

Review URL: https://codereview.chromium.org/23503046

git-svn-id: svn://svn.chromium.org/blink/trunk@159061 bbb929c8-8fbe-4397-9dbb-9b2b20218538