Don't output sessionID in encrypted-media-events.html
Previously, a sessionID was included in the logged output of a testExpected() statement, which caused text diffs when run multiple times (or in a different order). Change the way the test is logged so that the changing sessionID is not included in the output.

BUG=304991

Review URL: https://codereview.chromium.org/26405002

git-svn-id: svn://svn.chromium.org/blink/trunk@159079 bbb929c8-8fbe-4397-9dbb-9b2b20218538