Canonical name change: windows-949 => EUC-KR
This is to be compliant with the HTML5 encoding spec. The expected
result for encoding-labels.html was updated accordingly to pass EUC-KR
related tests.

It has to be noted that EUC-KR is a far more widely used encoding label than windows-949. Strictly speaking, Windows-949 is a superset of EUC-KR, but an widely accepted practice in the web is to treat them synonymously as codified in HTML 5. 

In addition, 'TextEncoding::domName' became unnecessary
(http://trac.webkit.org/changeset/43279) so that it's removed.

THe chromium side of change is https://codereview.chromium.org/26439002.


BUG=277027
TEST=fast/encoding/api/* and fast/encoding/euckr-name.html

Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=158972

Review URL: https://codereview.chromium.org/25870005

git-svn-id: svn://svn.chromium.org/blink/trunk@159125 bbb929c8-8fbe-4397-9dbb-9b2b20218538