Handle #READY lines from content_shell.
This is in preparation for making content_shell print out when it's
ready to start running tests so that we don't count content_shell
startup time as part of the timeout when running a test. Since
the C++ code in question is in the chromium tree, this needs
to be done as a 3-sided patch.

BUG=265088

Review URL: https://codereview.chromium.org/26571003

git-svn-id: svn://svn.chromium.org/blink/trunk@159216 bbb929c8-8fbe-4397-9dbb-9b2b20218538